requirements.txt
datasets>=2.16.0
huggingface-hub>=0.20.0
numpy>=1.24.0
openai>=1.12.0
pandas>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
tqdm>=4.66.0
typer>=0.9.0
typing-extensions>=4.8.0